Add tests for gpg tool.
Move in_path into the only place it is used.
