+++ lintreview/tools/gpg.py
@@ -1,12 +1,25 @@
 from __future__ import absolute_import
 import logging
+import os
 
 from lintreview.review import IssueComment
 from lintreview.tools import Tool
 from lintreview.tools import run_command
-from lintreview.utils import in_path
 
 log = logging.getLogger(__name__)
+
+
+def in_path(name):
+    """
+    Check whether or not a command line tool
+    exists in the system path.
+
+    @return boolean
+    """
+    for dirname in os.environ['PATH'].split(os.pathsep):
+        if os.path.exists(os.path.join(dirname, name)):
+            return True
+    return False
 
 
 class Gpg(Tool):
@@ -23,15 +36,12 @@ class Gpg(Tool):
         """
         Check that HEAD commit has gpg signature
         """
-        cmd = "git log HEAD^..HEAD --show-signature --format=%H | "
-        cmd += "grep -q 'Signature made'"
+        cmd = [
+            'git', 'log', 'HEAD^..HEAD',
+            '--show-signature', '--format=%H'
+        ]
 
-        try:
-            run_command(cmd, split=False, shell=True,
-                        ignore_error=False, cwd=self.base_path)
-            log.debug('Signature found in HEAD commit')
-            return False
-        except Exception as e:
-            log.debug("Exception: %s" % str(e))
+        output = run_command(cmd, ignore_error=True, cwd=self.base_path)
+        if 'Signature made' not in output:
             body = 'No gpg signature for tip of the branch.'
             self.problems.add(IssueComment(body))
